requirements.txt
fastapi
uvicorn
pillow
ultralytics
google-cloud-storage
